Reorganize tool namespaces & tests
