+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Commands/CommandRunner.cs
@@ -3,7 +3,7 @@ using System.CommandLine.Help;
 using Azure.Sdk.Tools.Cli.Commands.HostServer;
 using Azure.Sdk.Tools.Cli.Helpers;
 using Azure.Sdk.Tools.Cli.Telemetry;
-using Azure.Sdk.Tools.Cli.Tools;
+using Azure.Sdk.Tools.Cli.Tools.Core;
 
 namespace Azure.Sdk.Tools.Cli.Commands
 {
