+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Tools/EngSys/Cleanup/CleanupTool.cs
@@ -7,6 +7,7 @@ using Azure.Sdk.Tools.Cli.Services;
 using ModelContextProtocol.Server;
 using Azure.Sdk.Tools.Cli.Commands;
 using Azure.Sdk.Tools.Cli.Models;
+using Azure.Sdk.Tools.Cli.Tools.Core;
 
 namespace Azure.Sdk.Tools.Cli.Tools.EngSys;
