+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Tools/Example/HelloWorldTool.cs
@@ -6,6 +6,7 @@ using System.ComponentModel;
 using Azure.Sdk.Tools.Cli.Commands;
 using ModelContextProtocol.Server;
 using Azure.Sdk.Tools.Cli.Models;
+using Azure.Sdk.Tools.Cli.Tools.Core;
 
 namespace Azure.Sdk.Tools.Cli.Tools.Example
 {
